% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_sweep.R
\name{cf_benchmark_density_sweep}
\alias{cf_benchmark_density_sweep}
\title{Run confounder density sweep experiment}
\usage{
cf_benchmark_density_sweep(
  n_confounders_values = c(10, 50, 100, 200, 500),
  coef_scaling = "sqrt",
  n = 500,
  p = 500,
  n_reps = 10,
  seed = 123
)
}
\arguments{
\item{n_confounders_values}{Integer vector, number of confounders to test}

\item{coef_scaling}{Character, coefficient scaling method}

\item{n}{Integer, sample size}

\item{p}{Integer, number of covariates (must be >= max(n_confounders_values))}

\item{n_reps}{Integer, number of replications per setting}

\item{seed}{Integer, base random seed}
}
\value{
List of DGP results for each confounder count
}
\description{
Generates datasets with varying number of confounders.
}
