% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_diagnostics.R
\name{cf_cate_diagnostics}
\alias{cf_cate_diagnostics}
\title{CATE Diagnostics: Heterogeneity Validation}
\usage{
cf_cate_diagnostics(result, data, formula, n_groups = 4L)
}
\arguments{
\item{result}{A cfomics_result object}

\item{data}{Data frame used for fitting}

\item{formula}{Formula of the form Y ~ T | X1 + X2 + ...}

\item{n_groups}{Integer, number of ITE quantile groups for GATES (default 4)}
}
\value{
A list of class "cf_cate_diagnostics"
}
\description{
Tests for treatment effect heterogeneity using GATES (Sorted Group ATE)
and BLP (Best Linear Predictor) approaches from Chernozhukov et al. (2018).
}
