% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_traditional.R
\name{cf_fit_gformula}
\alias{cf_fit_gformula}
\title{Fit G-formula model}
\usage{
cf_fit_gformula(X, T, Y, covariate_names = NULL, nsimul = 1000L, ...)
}
\arguments{
\item{X}{Covariate matrix}

\item{T}{Treatment vector (binary)}

\item{Y}{Outcome vector}

\item{covariate_names}{Character vector of covariate names}

\item{nsimul}{Integer, number of simulations (default: 1000)}

\item{...}{Additional arguments}
}
\value{
List with model fit and results
}
\description{
Fit G-formula model
}
\keyword{internal}
