% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_traditional.R
\name{cf_fit_ipw}
\alias{cf_fit_ipw}
\title{Fit Inverse Probability Weighting (IPW) model}
\usage{
cf_fit_ipw(X, T, Y, covariate_names = NULL, stabilized = TRUE, ...)
}
\arguments{
\item{X}{Covariate matrix}

\item{T}{Treatment vector (binary)}

\item{Y}{Outcome vector}

\item{covariate_names}{Character vector of covariate names}

\item{stabilized}{Logical, whether to use stabilized weights (default: TRUE)}

\item{...}{Additional arguments passed to ipwpoint}
}
\value{
List with model fit and results
}
\description{
Fit Inverse Probability Weighting (IPW) model
}
\keyword{internal}
