% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/python_env.R
\name{cf_list_python_envs}
\alias{cf_list_python_envs}
\title{List available cfomics Python environments}
\usage{
cf_list_python_envs()
}
\value{
A data.frame with environment names and their status
}
\description{
List available cfomics Python environments
}
