% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf_predict.R
\name{cf_metadata}
\alias{cf_metadata}
\title{Get metadata from a cf_model}
\usage{
cf_metadata(object)
}
\arguments{
\item{object}{cf_model object}
}
\value{
A list containing metadata about the model fit, including
  input hash, DAG structure, random seed, and processing information.
}
\description{
Get metadata from a cf_model
}
