% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_diagnostics.R
\name{cf_model_diagnostics}
\alias{cf_model_diagnostics}
\title{Model Diagnostics for Causal Inference Results}
\usage{
cf_model_diagnostics(result, data, formula)
}
\arguments{
\item{result}{A cfomics_result object}

\item{data}{Data frame used for fitting}

\item{formula}{Formula of the form Y ~ T | X1 + X2 + ...}
}
\value{
A list of class "cf_model_diagnostics"
}
\description{
Evaluates the quality of propensity score and outcome models used in
causal inference estimation.
}
