% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_dgp.R
\name{dgp_dimension_sweep}
\alias{dgp_dimension_sweep}
\title{Generate high-dimensional DGP data (S2-S3 dimension sweep)}
\usage{
dgp_dimension_sweep(n = 200, p = 1000, seed = NULL)
}
\arguments{
\item{n}{Integer, number of observations}

\item{p}{Integer, number of covariates}

\item{seed}{Integer, random seed}
}
\value{
List with X, T, Y, true_ate, true_ite, propensity_score
}
\description{
Generates data with specified n and p to test methods under different
n/p ratios.
}
