% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_dgp.R
\name{dgp_heterogeneous_qualitative}
\alias{dgp_heterogeneous_qualitative}
\title{Generate qualitative interaction DGP (S4d)}
\usage{
dgp_heterogeneous_qualitative(n = 500, p = 500, seed = NULL)
}
\arguments{
\item{n}{Integer, number of observations}

\item{p}{Integer, number of covariates}

\item{seed}{Integer, random seed}
}
\value{
List with X, T, Y, true_ate, true_ite, prop_harmed
}
\description{
Treatment effect can be positive or negative depending on X1.
}
