% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_dgp.R
\name{dgp_heterogeneous_subgroup}
\alias{dgp_heterogeneous_subgroup}
\title{Generate subgroup heterogeneous treatment effect DGP (S4c/g/h)}
\usage{
dgp_heterogeneous_subgroup(
  n = 500,
  p = 500,
  subgroup_props = c(1/3, 1/3, 1/3),
  subgroup_effects = c(0.5, 2, 4),
  seed = NULL
)
}
\arguments{
\item{n}{Integer, number of observations}

\item{p}{Integer, number of covariates}

\item{subgroup_props}{Numeric vector, proportion in each subgroup}

\item{subgroup_effects}{Numeric vector, treatment effect in each subgroup}

\item{seed}{Integer, random seed}
}
\value{
List with X, T, Y, true_ate, true_ite, subgroup
}
\description{
Treatment effect varies by discrete subgroups.
}
