% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_dgp.R
\name{generate_benchmark_data}
\alias{generate_benchmark_data}
\title{Generate benchmark data by DGP name}
\usage{
generate_benchmark_data(dgp, params = list())
}
\arguments{
\item{dgp}{Character, the DGP name (e.g., "baseline", "nonlinear_outcome")}

\item{params}{List of parameters passed to the DGP function}
}
\value{
List with X, T, Y, true_ate, true_ite, etc.
}
\description{
Dispatches to the appropriate DGP function based on a string name.
}
