% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot.cf_balance}
\alias{plot.cf_balance}
\title{Plot Covariate Balance}
\usage{
\method{plot}{cf_balance}(x, ...)
}
\arguments{
\item{x}{cf_balance object}

\item{...}{Additional arguments passed to plot}
}
\description{
Plot Covariate Balance
}
