% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_diagnostics.R
\name{plot.cf_cate_diagnostics}
\alias{plot.cf_cate_diagnostics}
\title{Plot CATE diagnostics}
\usage{
\method{plot}{cf_cate_diagnostics}(x, ...)
}
\arguments{
\item{x}{A cf_cate_diagnostics object}

\item{...}{Additional arguments (ignored)}
}
\description{
Plot CATE diagnostics
}
