% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{plot.cf_overlap}
\alias{plot.cf_overlap}
\title{Plot Overlap Diagnostics}
\usage{
\method{plot}{cf_overlap}(x, ...)
}
\arguments{
\item{x}{cf_overlap object}

\item{...}{Additional arguments}
}
\description{
Plot Overlap Diagnostics
}
