% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_cavae.R
\name{predict_cf_cavae}
\alias{predict_cf_cavae}
\title{Predict from CAVAE model}
\usage{
predict_cf_cavae(object, newdata = NULL, type = "ite", ...)
}
\arguments{
\item{object}{cf_model object with method="cavae"}

\item{newdata}{Optional new data for prediction}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1"}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict from CAVAE model
}
\keyword{internal}
