% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_drlearner.R
\name{predict_cf_drlearner}
\alias{predict_cf_drlearner}
\title{Predict from DRLearner model}
\usage{
predict_cf_drlearner(object, newdata = NULL, type = "ite", ...)
}
\arguments{
\item{object}{cf_model object with method="drlearner"}

\item{newdata}{Optional new data for prediction}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1"}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict from DRLearner model
}
\keyword{internal}
