% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_ganite.R
\name{predict_cf_ganite}
\alias{predict_cf_ganite}
\title{Predict from GANITE model}
\usage{
predict_cf_ganite(
  object,
  type = c("ate", "ite", "y0", "y1", "summary", "samples"),
  ...
)
}
\arguments{
\item{object}{cf_model object with method="ganite"}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1", "summary", "samples"}

\item{...}{Additional arguments (unused)}
}
\value{
Predicted values
}
\description{
Predict from GANITE model
}
\keyword{internal}
