% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_traditional.R
\name{predict_cf_gformula}
\alias{predict_cf_gformula}
\title{Predict from G-formula model}
\usage{
predict_cf_gformula(object, newdata = NULL, type = "ite", ...)
}
\arguments{
\item{object}{cf_model object with method="gformula"}

\item{newdata}{Optional new data for prediction}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1"}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict from G-formula model
}
\keyword{internal}
