% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hdml.R
\name{predict_cf_hdml}
\alias{predict_cf_hdml}
\title{Predict from HDML model}
\usage{
predict_cf_hdml(object, newdata = NULL, type = "ite", ...)
}
\arguments{
\item{object}{cf_model object with method="hdml"}

\item{newdata}{Optional new data for prediction}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1"}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict from HDML model
}
\keyword{internal}
