% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_hdps.R
\name{predict_cf_hdps}
\alias{predict_cf_hdps}
\title{Predict from HDPS model}
\usage{
predict_cf_hdps(object, newdata = NULL, type = "ite", ...)
}
\arguments{
\item{object}{cf_model object with method="hdps"}

\item{newdata}{Not supported for HDPS}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1"}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict from HDPS model
}
\keyword{internal}
