% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_traditional.R
\name{predict_cf_ipw}
\alias{predict_cf_ipw}
\title{Predict from IPW model}
\usage{
predict_cf_ipw(object, newdata = NULL, type = "ite", ...)
}
\arguments{
\item{object}{cf_model object with method="ipw"}

\item{newdata}{Not supported for IPW}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1"}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict from IPW model
}
\keyword{internal}
