% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_tmle.R
\name{predict_cf_tmle}
\alias{predict_cf_tmle}
\title{Predict from TMLE model}
\usage{
predict_cf_tmle(object, newdata = NULL, type = "ite", ...)
}
\arguments{
\item{object}{cf_model object with method="tmle"}

\item{newdata}{Optional new data for prediction}

\item{type}{Type of prediction: "ate", "ite", "y0", "y1", "summary", "samples"}

\item{...}{Additional arguments}
}
\value{
Predicted values
}
\description{
Predict from TMLE model
}
\keyword{internal}
