% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.cf_balance}
\alias{print.cf_balance}
\title{Print method for cf_balance}
\usage{
\method{print}{cf_balance}(x, ...)
}
\arguments{
\item{x}{cf_balance object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for cf_balance
}
