% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cate_diagnostics.R
\name{print.cf_cate_diagnostics}
\alias{print.cf_cate_diagnostics}
\title{Print CATE diagnostics}
\usage{
\method{print}{cf_cate_diagnostics}(x, ...)
}
\arguments{
\item{x}{A cf_cate_diagnostics object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print CATE diagnostics
}
