% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{print.cf_overlap}
\alias{print.cf_overlap}
\title{Print method for cf_overlap}
\usage{
\method{print}{cf_overlap}(x, ...)
}
\arguments{
\item{x}{cf_overlap object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for cf_overlap
}
