% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{print.cf_session_info}
\alias{print.cf_session_info}
\title{Print method for cf_session_info}
\usage{
\method{print}{cf_session_info}(x, ...)
}
\arguments{
\item{x}{cf_session_info object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for cf_session_info
}
