% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_print.R
\name{print.cfomics_result}
\alias{print.cfomics_result}
\title{Print method for cfomics_result}
\usage{
\method{print}{cfomics_result}(x, ...)
}
\arguments{
\item{x}{A cfomics_result object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for cfomics_result
}
